+++ k256/src/arithmetic/affine.rs
@@ -1,7 +1,7 @@
 //! Affine points
 
-use super::{FieldElement, ProjectivePoint, CURVE_EQUATION_A, CURVE_EQUATION_B, MODULUS};
-use crate::{EncodedPoint, FieldBytes, NistP256, NonZeroScalar};
+use super::{FieldElement, ProjectivePoint, CURVE_EQUATION_B};
+use crate::{EncodedPoint, FieldBytes, NonZeroScalar, Secp256k1};
 use core::ops::{Mul, Neg};
 use elliptic_curve::{
     generic_array::arr,
@@ -13,7 +13,7 @@ use elliptic_curve::{
 #[cfg(feature = "zeroize")]
 use elliptic_curve::zeroize::Zeroize;
 
-/// A point on the secp256r1 curve in affine coordinates.
+/// A point on the secp256k1 curve in affine coordinates.
 #[derive(Clone, Copy, Debug)]
 #[cfg_attr(docsrs, doc(cfg(feature = "arithmetic")))]
 pub struct AffinePoint {
@@ -23,22 +23,22 @@ pub struct AffinePoint {
 }
 
 impl AffinePoint {
-    /// Returns the base point of P-256.
+    /// Returns the base point of SECP256k1.
     pub fn generator() -> AffinePoint {
-        // NIST P-256 basepoint in affine coordinates:
-        // x = 6b17d1f2 e12c4247 f8bce6e5 63a440f2 77037d81 2deb33a0 f4a13945 d898c296
-        // y = 4fe342e2 fe1a7f9b 8ee7eb4a 7c0f9e16 2bce3357 6b315ece cbb64068 37bf51f5
+        // SECP256k1 basepoint in affine coordinates:
+        // x = 79be667e f9dcbbac 55a06295 ce870b07 029bfcdb 2dce28d9 59f2815b 16f81798
+        // y = 483ada77 26a3c465 5da4fbfc 0e1108a8 fd17b448 a6855419 9c47d08f fb10d4b8
         AffinePoint {
             x: FieldElement::from_bytes(&arr![u8;
-                0x6b, 0x17, 0xd1, 0xf2, 0xe1, 0x2c, 0x42, 0x47, 0xf8, 0xbc, 0xe6, 0xe5, 0x63, 0xa4,
-                0x40, 0xf2, 0x77, 0x03, 0x7d, 0x81, 0x2d, 0xeb, 0x33, 0xa0, 0xf4, 0xa1, 0x39, 0x45,
-                0xd8, 0x98, 0xc2, 0x96
+                0x79, 0xbe, 0x66, 0x7e, 0xf9, 0xdc, 0xbb, 0xac, 0x55, 0xa0, 0x62, 0x95, 0xce, 0x87,
+                0x0b, 0x07, 0x02, 0x9b, 0xfc, 0xdb, 0x2d, 0xce, 0x28, 0xd9, 0x59, 0xf2, 0x81, 0x5b,
+                0x16, 0xf8, 0x17, 0x98
             ])
             .unwrap(),
             y: FieldElement::from_bytes(&arr![u8;
-                0x4f, 0xe3, 0x42, 0xe2, 0xfe, 0x1a, 0x7f, 0x9b, 0x8e, 0xe7, 0xeb, 0x4a, 0x7c, 0x0f,
-                0x9e, 0x16, 0x2b, 0xce, 0x33, 0x57, 0x6b, 0x31, 0x5e, 0xce, 0xcb, 0xb6, 0x40, 0x68,
-                0x37, 0xbf, 0x51, 0xf5
+                0x48, 0x3a, 0xda, 0x77, 0x26, 0xa3, 0xc4, 0x65, 0x5d, 0xa4, 0xfb, 0xfc, 0x0e, 0x11,
+                0x08, 0xa8, 0xfd, 0x17, 0xb4, 0x48, 0xa6, 0x85, 0x54, 0x19, 0x9c, 0x47, 0xd0, 0x8f,
+                0xfb, 0x10, 0xd4, 0xb8
             ])
             .unwrap(),
             infinity: Choice::from(0),
@@ -72,7 +72,9 @@ impl ConditionallySelectable for AffinePoint {
 
 impl ConstantTimeEq for AffinePoint {
     fn ct_eq(&self, other: &AffinePoint) -> Choice {
-        self.x.ct_eq(&other.x) & self.y.ct_eq(&other.y) & self.infinity.ct_eq(&other.infinity)
+        (self.x.negate(1) + &other.x).normalizes_to_zero()
+            & (self.y.negate(1) + &other.y).normalizes_to_zero()
+            & self.infinity.ct_eq(&other.infinity)
     }
 }
 
@@ -90,22 +92,22 @@ impl PartialEq for AffinePoint {
 
 impl Eq for AffinePoint {}
 
-impl Decompress<NistP256> for AffinePoint {
+impl Decompress<Secp256k1> for AffinePoint {
     fn decompress(x_bytes: &FieldBytes, y_is_odd: Choice) -> CtOption<Self> {
         FieldElement::from_bytes(x_bytes).and_then(|x| {
-            let alpha = x * &x * &x + &(CURVE_EQUATION_A * &x) + &CURVE_EQUATION_B;
+            let alpha = (x * &x * &x) + &CURVE_EQUATION_B;
             let beta = alpha.sqrt();
 
             beta.map(|beta| {
                 let y = FieldElement::conditional_select(
-                    &(MODULUS - &beta),
+                    &beta.negate(1),
                     &beta,
-                    beta.is_odd().ct_eq(&y_is_odd),
+                    beta.normalize().is_odd().ct_eq(&y_is_odd),
                 );
 
                 Self {
                     x,
-                    y,
+                    y: y.normalize(),
                     infinity: Choice::from(0),
                 }
             })
@@ -113,12 +115,12 @@ impl Decompress<NistP256> for AffinePoint {
     }
 }
 
-impl FromEncodedPoint<NistP256> for AffinePoint {
+impl FromEncodedPoint<Secp256k1> for AffinePoint {
     /// Attempts to parse the given [`EncodedPoint`] as an SEC1-encoded [`AffinePoint`].
     ///
     /// # Returns
     ///
-    /// `None` value if `encoded_point` is not on the secp256r1 curve.
+    /// `None` value if `encoded_point` is not on the secp256k1 curve.
     fn from_encoded_point(encoded_point: &EncodedPoint) -> CtOption<Self> {
         match encoded_point.coordinates() {
             sec1::Coordinates::Compressed { x, y_is_odd } => {
@@ -131,14 +133,14 @@ impl FromEncodedPoint<NistP256> for AffinePoint {
                 x.and_then(|x| {
                     y.and_then(|y| {
                         // Check that the point is on the curve
-                        let lhs = y * &y;
-                        let rhs = x * &x * &x + &(CURVE_EQUATION_A * &x) + &CURVE_EQUATION_B;
+                        let lhs = (y * &y).negate(1);
+                        let rhs = x * &x * &x + &CURVE_EQUATION_B;
                         let point = AffinePoint {
                             x,
                             y,
                             infinity: Choice::from(0),
                         };
-                        CtOption::new(point, lhs.ct_eq(&rhs))
+                        CtOption::new(point, (lhs + &rhs).normalizes_to_zero())
                     })
                 })
             }
@@ -146,7 +148,7 @@ impl FromEncodedPoint<NistP256> for AffinePoint {
     }
 }
 
-impl ToEncodedPoint<NistP256> for AffinePoint {
+impl ToEncodedPoint<Secp256k1> for AffinePoint {
     fn to_encoded_point(&self, compress: bool) -> EncodedPoint {
         EncodedPoint::from_affine_coordinates(&self.x.to_bytes(), &self.y.to_bytes(), compress)
     }
@@ -155,7 +157,7 @@ impl ToEncodedPoint<NistP256> for AffinePoint {
 impl From<AffinePoint> for EncodedPoint {
     /// Returns the SEC1 compressed encoding of this point.
     fn from(affine_point: AffinePoint) -> EncodedPoint {
-        affine_point.to_encoded_point(false)
+        affine_point.to_encoded_point(true)
     }
 }
 
@@ -173,7 +175,7 @@ impl Neg for AffinePoint {
     fn neg(self) -> Self::Output {
         AffinePoint {
             x: self.x,
-            y: -self.y,
+            y: self.y.negate(1).normalize_weak(),
             infinity: self.infinity,
         }
     }
@@ -195,30 +197,29 @@ mod tests {
     use hex_literal::hex;
 
     const UNCOMPRESSED_BASEPOINT: &[u8] = &hex!(
-        "046B17D1F2E12C4247F8BCE6E563A440F277037D812DEB33A0F4A13945D898C296
-         4FE342E2FE1A7F9B8EE7EB4A7C0F9E162BCE33576B315ECECBB6406837BF51F5"
+        "0479BE667EF9DCBBAC55A06295CE870B07029BFCDB2DCE28D959F2815B16F81798
+         483ADA7726A3C4655DA4FBFC0E1108A8FD17B448A68554199C47D08FFB10D4B8"
     );
-
     const COMPRESSED_BASEPOINT: &[u8] =
-        &hex!("036B17D1F2E12C4247F8BCE6E563A440F277037D812DEB33A0F4A13945D898C296");
+        &hex!("0279BE667EF9DCBBAC55A06295CE870B07029BFCDB2DCE28D959F2815B16F81798");
 
     #[test]
     fn uncompressed_round_trip() {
         let pubkey = EncodedPoint::from_bytes(UNCOMPRESSED_BASEPOINT).unwrap();
-        let point = AffinePoint::from_encoded_point(&pubkey).unwrap();
-        assert_eq!(point, AffinePoint::generator());
+        let res: EncodedPoint = AffinePoint::from_encoded_point(&pubkey)
+            .unwrap()
+            .to_encoded_point(false);
 
-        let res: EncodedPoint = point.into();
         assert_eq!(res, pubkey);
     }
 
     #[test]
     fn compressed_round_trip() {
         let pubkey = EncodedPoint::from_bytes(COMPRESSED_BASEPOINT).unwrap();
-        let point = AffinePoint::from_encoded_point(&pubkey).unwrap();
-        assert_eq!(point, AffinePoint::generator());
+        let res: EncodedPoint = AffinePoint::from_encoded_point(&pubkey)
+            .unwrap()
+            .to_encoded_point(true);
 
-        let res: EncodedPoint = point.to_encoded_point(true);
         assert_eq!(res, pubkey);
     }
 
@@ -254,6 +255,6 @@ mod tests {
     #[test]
     fn affine_negation() {
         let basepoint = AffinePoint::generator();
-        assert_eq!(-(-basepoint), basepoint);
+        assert_eq!((-(-basepoint)), basepoint);
     }
 }